+++ tests/functional/a/access/access_member_before_definition.py
@@ -1,11 +1,12 @@
 # pylint: disable=missing-docstring,too-few-public-methods,invalid-name
 # pylint: disable=attribute-defined-outside-init, useless-object-inheritance
 
+
 class Aaaa(object):
     """class with attributes defined in wrong order"""
 
     def __init__(self):
-        var1 = self._var2 # [access-member-before-definition]
+        var1 = self._var2  # [access-member-before-definition]
         self._var2 = 3
         self._var3 = var1
 
@@ -21,20 +22,26 @@ class Bbbb(object):
             self.__repo = attr
             return attr
 
-
     def catchme(self, attr):
         """no AttributeError caught"""
         try:
-            return self._repo # [access-member-before-definition]
+            return self._repo  # [access-member-before-definition]
         except ValueError:
             self._repo = attr
             return attr
 
 
 class Mixin(object):
-
     def test_mixin(self):
         """Don't emit access-member-before-definition for mixin classes."""
         if self.already_defined:
             # pylint: disable=attribute-defined-outside-init
             self.already_defined = None
+
+
+# Test for regression in bitbucket issue 164
+# https://bitbucket.org/logilab/pylint/issue/164/
+class MyClass1:
+    def __init__(self):
+        self.first += 5  # [access-member-before-definition]
+        self.first = 0
