+++ tests/functional/a/access/access_member_before_definition.txt
@@ -1,2 +1,3 @@
-access-member-before-definition:8:15:8:25:Aaaa.__init__:Access to member '_var2' before its definition line 9:UNDEFINED
+access-member-before-definition:9:15:9:25:Aaaa.__init__:Access to member '_var2' before its definition line 10:UNDEFINED
 access-member-before-definition:28:19:28:29:Bbbb.catchme:Access to member '_repo' before its definition line 30:UNDEFINED
+access-member-before-definition:46:8:46:18:MyClass1.__init__:Access to member 'first' before its definition line 47:UNDEFINED
