+++ tests/functional/m/method_hidden.py
@@ -1,30 +1,38 @@
 # pylint: disable=too-few-public-methods, useless-object-inheritance,missing-docstring
+# pylint: disable=unused-private-member
 """check method hidding ancestor attribute
 """
 from __future__ import print_function
 
+
 class Abcd(object):
     """dummy"""
+
     def __init__(self):
         self.abcd = 1
 
+
 class Cdef(Abcd):
     """dummy"""
-    def abcd(self): # [method-hidden]
-        """test
-        """
+
+    def abcd(self):  # [method-hidden]
+        """test"""
         print(self)
+
 
 class AbcdMixin:
     def abcd(self):
         pass
 
+
 class Dabc(AbcdMixin, Abcd):
     def abcd(self):
         pass
 
+
 class CustomProperty:
     """dummy"""
+
     def __init__(self, _):
         pass
 
@@ -36,8 +44,10 @@ class CustomProperty:
     def __set__(self, _, __):
         pass
 
+
 class Ddef:
     """dummy"""
+
     def __init__(self):
         self.five = "five"
 
@@ -73,12 +83,14 @@ class One:
         if one is not None:
             self.one = one
 
-    def one(self): # [method-hidden]
+    def one(self):  # [method-hidden]
         pass
+
 
 class Two(One):
     def one(self):
         pass
+
 
 try:
     import unknown as js
@@ -90,3 +102,23 @@ class JsonEncoder(js.JSONEncoder):
     # pylint: disable=useless-super-delegation,super-with-arguments
     def default(self, o):
         return super(JsonEncoder, self).default(o)
+
+
+class Parent:
+    def __init__(self):
+        self._protected = None
+
+
+class Child(Parent):
+    def _protected(self):  # [method-hidden]
+        pass
+
+
+class ParentTwo:
+    def __init__(self):
+        self.__private = None
+
+
+class ChildTwo(ParentTwo):
+    def __private(self):
+        pass
