+++ tests/functional/p/protected_access_special_methods_off.rc
@@ -1,0 +1,2 @@
+[MASTER]
+check-protected-access-in-special-methods=no
